Documentation update in features.py, minor optimization added to phase 1 selection loop in FeatureSelector.choose(). Also bringing unit_tests_features up to speed.
